Use Common Base Class for Cubes
